tourist

+++ api/serializers.py
@@ -5,5 +5,5 @@ class UserInputSerializer(serializers.Serializer):
     budget = serializers.IntegerField()
     available_from = serializers.CharField()
     available_to = serializers.CharField()
-    lat = serializers.IntegerField()
-    long = serializers.IntegerField()
+    lat = serializers.FloatField()
+    long = serializers.FloatField()
